moved functions to separate file
