+++ assets/js/constants.js
@@ -1,1 +1,6 @@
 export const CLICK_EVENT = "click";
+export const icons = {
+    EDIT: `<i class="fa-solid fa-pen" id="edit"></i>`,
+    DELETE: `<i class="fa-solid fa-trash-can" id="delete"></i>`,
+    DONE: `<i class="fa-solid fa-check" id="done"></i>`,
+  };
