+++ assets/js/main.js
@@ -8,11 +8,6 @@ window.addEventListener("load", () => {
   const cancelButton = document.querySelector("#cancel-button");
   const allTasks = {};
   let id = 0;
-  const icons = {
-    EDIT: `<i class="fa-solid fa-pen" id="edit"></i>`,
-    DELETE: `<i class="fa-solid fa-trash-can" id="delete"></i>`,
-    DONE: `<i class="fa-solid fa-check" id="done"></i>`,
-  };
   addTaskEvent(
     createButton,
     cancelButton,
@@ -20,7 +15,6 @@ window.addEventListener("load", () => {
     inputValue,
     allTasks,
     id,
-    taskList,
-    icons
+    taskList
   );
 });
